requirements.txt
streamlit
pandas
python-docx
langchain-text-splitters
python-dotenv
chromadb
groq
sentence-transformers
tabulate
openpyxl
